Update README
Motivation:

After moving to GitLab, and while still having a GitHub mirror, it is important to drive people to the proper location for getting help and reporting bugs or feature requests.

Also, some information should be updated to include new tags, formatting, and a link to the new API docs.

Modifications:

README is updated to point users to the GitLab repository and to have up-to-date information about the project.

Result:

Users should know where to reach project maintainers and where to find documentation.