Make `RedisCommandHandler` final
Motivation:

During SSWG review, feedback was provided that forced a re-evaluation of early design feedback interpretation on composability of RedisNIO.

First understanding was that the desire was to have "customization points" to gain benefits of implementation of RedisCommandHandler, while new understanding is that it just needs to be public in order for users to include it in their own custom ChannelPipeline schemes.

Modifications:

`RedisCommandHandler` is now a final class.

Result:

Users will no longer be able to subclass `RedisCommandHandler`, but gain a super slight performance increase.

This contributes to #47.