Speed up CI pipelines for Merge Requests
Motivation:

MRs take unnecessarily too long to have completed pipelines with "redundant" builds happening in sequence.

The entire stage of building in release is an additional QA layer that can happen on merges into master, but for MRs are just added CPU cycles.

While it's useful to ensure that building in release is possible, we will get almost all build failures from the test stage, and the most important checks are the unit and integration tests.

Modifications:

Change the ordering of build vs. test, as well as only running "build in release" jobs on the master branch.

Result:

MRs should be able to be merged sooner as CI pipelines take less time to run.