Convert `RESPTranslator.ParsingError` from enum to struct
Motivation:

Much like the change for `RedisClientError` (6471a2) any error enum that we may want to change will necessitate a major SemVer.

To avoid this, we need to use the established struct-as-enum pattern.

Modifications:

- Change: `RESPTranslator.ParsingError` to be a struct backed by an enum

Result:

The library should be able to evolve to add more `RESPTranslator.ParsingError` cases without becoming a breaking change.